refactor(manufacture): extract FlexiIngredientRequirementAggregator
