+++ backend/src/Adapters/Anela.Heblo.Adapters.Flexi/Manufacture/Internal/IFlexiIngredientRequirementAggregator.cs
@@ -1,0 +1,10 @@
+using Anela.Heblo.Domain.Features.Manufacture;
+
+namespace Anela.Heblo.Adapters.Flexi.Manufacture.Internal;
+
+internal interface IFlexiIngredientRequirementAggregator
+{
+    Task<Dictionary<string, IngredientRequirement>> AggregateAsync(
+        IReadOnlyList<SubmitManufactureClientItem> items,
+        CancellationToken cancellationToken);
+}
